fix(davinci): split css_bind authored-span helpers under the length gate
The exact provenance pin pushed css_bind_sfc.rs over the 350-line
ceiling; move the span-validation helpers into an authored module of a
directory-style test target (the metamorphic lane's layout).
